--- a/daemon/container/view_test.go
+++ b/daemon/container/view_test.go
@@ -30,7 +30,7 @@ func TestViewSaveDelete(t *testing.T) {
 	tmpDir := t.TempDir()
 	c := newContainer(t, tmpDir)
 	assert.NilError(t, c.CheckpointTo(context.Background(), db))
-	db.Delete(c)
+	db.Delete(c.ID)
 }
 
 func TestViewAll(t *testing.T) {
@@ -130,7 +130,7 @@ func TestNames(t *testing.T) {
 	assert.Check(t, is.DeepEqual(map[string][]string{"containerid1": {"name1", "name3", "name4"}, "containerid4": {"name2"}}, view.GetAllNames()))
 
 	// Release containerid1's names with Delete even though no container exists
-	db.Delete(&Container{ID: "containerid1"})
+	db.Delete("containerid1")
 
 	// Reusing one of those names should work
 	assert.Check(t, db.ReserveName("name1", "containerid4"))
@@ -271,7 +271,7 @@ func TestTruncIndex(t *testing.T) {
 	}
 
 	// Deleting id2 should remove conflicts
-	db.Delete(&Container{ID: id2})
+	db.Delete(id2)
 
 	for _, tc := range []testacase{
 		{
@@ -466,7 +466,7 @@ func BenchmarkDBDelete(b *testing.B) {
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		for _, id := range testKeys {
-			db.Delete(&Container{ID: id})
+			db.Delete(id)
 		}
 	}
 }