### container: ViewDB.Delete: remove error return

This function never returns an error, so let's remove the return.


### container: add BenchmarkDBDelete


    go test -bench=BenchmarkDBDelete -run=^#
    goos: linux
    goarch: arm64
    pkg: github.com/docker/docker/container
    BenchmarkDBdDelete-10    	     172	   7322767 ns/op	10626564 B/op	  105000 allocs/op
    PASS
    ok  	github.com/docker/docker/container	1.819s


### container: ViewDB.Delete: accept container ID, and optimization

This started because I was working on some changes in NewBaseContainer,
when I noticed that `ViewDb.Delete()` accepted a Container struct as argument,
but then constructed a _new_ Container struct, preserving only the original
container's ID and Root. I noticed that for other parts of the code, we define
a dedicated type, but for Containers, we kept the same to keep it symmetrical,
and because `Save()` expects a container.

Looking at the code, effectively all that's needed is the ID (which is used
as primary key), and for `FromObject` to resolve it from an object that's
passed (for `Save()` that would be a `Container` struct).

So as a fun exercise, I modified `FromObject` to accept either a `Container`
(for the `Save()` case) or a new `containerIDAssociation` type (a string),
so that we don't have to construct a new `Container` just for sake of being
able to extract the ID.

As we no longer need a `Container` struct for this, I also changed the signature
of `ViewDB.Delete` to accept a string,

I don't expect `Delete()` to be in a very "hot path", but here's benchmarks
before and after;

    go test -bench=BenchmarkDBDelete -run=^#
    BenchmarkDBdDelete-10    	     172	   7322767 ns/op	10626564 B/op	  105000 allocs/op

    go test -bench=BenchmarkDBDelete -run=^#
    BenchmarkDBDelete-10    	     261	   4596451 ns/op	 4546504 B/op	   80000 allocs/op

And diff using `benchstat`:


|             | sec/op    | before       | after         | diff                   |
|-------------|-----------|--------------|---------------|------------------------|
| DBDelete-10 | sec/op    | 6.305m ± 3%  | 4.587m ± 8%   | -27.25% (p=0.000 n=10) |
| DBDelete-10 | B/op      | 7.006Mi ± 0% | 4.336Mi ± 0%  | -38.11% (p=0.000 n=10) |
| DBDelete-10 | allocs/op | 100.00k ± 0% | 80.00k ± 0%   | -20.00% (p=0.000 n=10) |




**- Description for the changelog**
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog.
It must be placed inside the below triple backticks section:
-->
```markdown changelog

```

**- A picture of a cute animal (not mandatory but encouraged)**