@thaJeztah 2025-08-21
Ah, lol, looks like I dropped a snippet in my previous rebase. I was scratching my head because it was green before, and last rebase broken that; this snipped was missing (see commit before previous rebase https://github.com/thaJeztah/docker/commit/733f057cb39c7909039b02968d443b1415372035), which 🙈 definitely explains why CI wasn't happy 😂 



```patch
diff --git a/daemon/container/view.go b/daemon/container/view.go
index 2105dee657..099e017a22 100644
--- a/daemon/container/view.go
+++ b/daemon/container/view.go
@@ -451,12 +451,16 @@ type containerByIDIndexer struct{}

 // FromObject implements the memdb.SingleIndexer interface for Container objects
 func (e *containerByIDIndexer) FromObject(obj any) (bool, []byte, error) {
-       c, ok := obj.(*Container)
-       if !ok {
+       switch c := obj.(type) {
+       case containerIDKey:
+               // Add the null character as a terminator
+               return true, []byte(c + terminator), nil
+       case *Container:
+               // Add the null character as a terminator
+               return true, []byte(c.ID + terminator), nil
+       default:
                return false, nil, fmt.Errorf("%T is not a Container", obj)
        }
```